move some hint generating logic outside of slice
